LayoutTests/editing/execCommand/format-block-table-expected.txt


Formatting:
| <#selection-anchor>
| "\n"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         "hello"
|       <td>
|         "world"
| "\n"
| <#selection-focus>

by p yields:
| "\n"
| <p>
|   <#selection-anchor>
|   <table>
|     <tbody>
|       <tr>
|         <td>
|           "hello"
|         <td>
|           "world"
|   <#selection-focus>
| "\n"

Formatting:
| "\n"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <#selection-anchor>
|         "hello"
|         <#selection-focus>
|       <td>
|         "world"
| "\n"

by blockquote yields:
| "\n"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <blockquote>
|           "<#selection-anchor>hello<#selection-focus>"
|       <td>
|         "world"
| "\n"

Formatting:
| "\n"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <#selection-anchor>
|         <p>
|           "hello"
|         <div>
|           "world"
|         <#selection-focus>
|       <td>
|         "WebKit"
| "\n"

by h3 yields:
| "\n"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <h3>
|           "<#selection-anchor>hello"
|           <br>
|           "world<#selection-focus>"
|       <td>
|         "WebKit"
| "\n"

Formatting:
| "\n"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <#selection-anchor>
|         <ul>
|           <li>
|             "hello"
|           <li>
|             "world"
|         <#selection-focus>
|       <td>
|         "WebKit"
| "\n"

by address yields:
| "\n"
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <address>
|           <ul>
|             <li>
|               "<#selection-anchor>hello"
|           <ul>
|             <li>
|               "world<#selection-focus>"
|       <td>
|         "WebKit"
| "\n"
